Renamed mock file of Number component.
